Update batchreadDPE128x64.m
Replaced all of the " if n == " statements (lines 50-64) with a single "repmat" (line 49).
Two reasons: First, makes it shorter/easier to read (for me), and second, I think we will want to be able to read subsections of the array for arbitrary n. However, it looks like this code will need more changes before we can read an arbitrary subarray, so let's discuss it tomorrow.